Remove check for charm config being set to same value. Bug #2043613.
This check was added 10 years ago to fix bug #1384622. The behaviour
back then was different, setting to the same value would have no effect.
Now, if the defaults have changed and it has been set the value will not
change, otherwise it will. The error message implied juju performed a
no-op but this is not the case.