feat(cloud): Add GetIDForCloud function
Add a function to get the uuid of a cloud. Currently clouds only have a
uuid in domain.

The cloud uuid should be used in the grant_on field of permissions
rather than the cloud name. The issue with this is that the core cloud
type does not have the UUID in it.

Rather than piping cloud UUIDs through the whole of Juju right now we
have instead decided to add this method to the cloud domain to retrive
the cloud UUID when needed.

The idea is that this method can be replaced once the cloud UUID is part
of the cloud definition.