feat: resource store type
The service doesn't need to know about the object, it just cares
that given a resource type, we get a store and using that store, we
can then get a io.Reader. Talking about narrow types prevents us
from using general types. This aids with testing and generalization.

We've removed coupling from a general type.