export and import using resource description v2
Add resource export and import using the new version of the description
package. This exports application and unit resources, kubernetes
application resources are not yet export, this will be done in a future
PR.

This PR uses the new version of the description package, tested by using
a replace in the go.mod.