fix: only get available resources when selection resource UUID.
GetApplicationResourceID and GetResourceUUIDByApplicationAndResourceName
could return potential resources as well as available resources. This is
not useful for any of their callers and has potential to cause confusing
bugs.

This was caught in migration where we would try to migrate a potentail
resource thinking it was available. The select statement was missing a
filter on using available resources only.