fix: return size and fingerprint from resource store
When a container image resource is downloaded from charmhub or uploaded
from the CLI, it can either be stored as json or yaml in a blob. When
the resource is fetched from the container image resource store, the
data is re-searialized into a json blob. This blob may be different to
the original blob Put into to the store, since the format may be
different, or the json may be serealised differently. This means the
size/fingerprint for the blob put into the store may not match the
size/fingerprint of the blob read out of the store.

This bug was disocvered in 4.0-4.0 migrations where the size in the
header of a container image resource being sent over http did not match
the size of the body being written.

The size matched the original blob size downloaded from charmhub, but
the blob being written was different. Specifically, these were the bytes
downloaded from charmhub:
{"ImageName": "registry.hub.docker.com/benhoyt/snappass-test..."}
And these were the bytes of the blob being written:
{"ImageName":"registry.hub.docker.com/benhoyt/snappass-test...","username":"","password":""}

To fix this issue of mismatching size/fingerprints, rather than
recording the size/fingerprint of the blob from charmhub/the cli in the
database, the size/fingerprint are returned from the resource store, and
these are stored in the database. This way we can get the correct values
that are controlled by the store, not by an external service.

As a flyby, a logger is added to the containerimageresourcestore.